pyproject.toml
[tool.poetry]
name = "eelapp"
version = "0.1.0"
description = ""
authors = ["Anthony Keelan <anthony.keelan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
eel = "^0.11.0"
pyinstaller = "^3.5"
sqlalchemy = "^1.3"
flask-sqlalchemy = "^2.4"
marshmallow = "^3.2"
marshmallow-sqlalchemy = "^0.19.0"
flask-marshmallow = "^0.10.1"

[tool.poetry.dev-dependencies]
yapf = "^0.28.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
